Add WP snippet
This snippet must be added anywhere on the source WordPress site. It will bake global styles into a static file under /wp-content/ directory so we can refer to it later on our nextjs-based blog.

The static file will be updated on every load of the main Dashboard screen (when a user is logged-in).